api/web/requirements.txt
PyMySQL
Flask==1.1.4
markupsafe==2.0.1
python-dotenv
PyJWT